EmbeddedModels/gemini_embeddings.py
from langchain_google_genai import GoogleGenerativeAIEmbeddings
from dotenv import load_dotenv

load_dotenv()

embeddings = GoogleGenerativeAIEmbeddings(model="models/gemini-embedding-exp-03-07")
res=embeddings.embed_query("Captial of India?")
print(str(res))